Completed more dart files
Finished up code for api and firebase calls, as well as the add exercise screen. Left exercise_screen.dart to do, streambuilder is in progress so i commented it out for the time being. Also should i add "a second brain" for the application? It feels kinda dumb without it